Update README with upgrade instructions and build steps
Introduction of new sections "Upgrade" and the inclusion of build and execute steps in the "Usage" section of the README file. It now provides detailed instructions⚙️ on how to upgrade to the latest version of AutoCommit and on building and executing with command line.

- A new section "Upgrade" 📈 was added to guide the user on how to upgrade to the latest version of AutoCommit. This serves to ensure users have the most recent features and fixes, optimizing their user experience.

- Under the "Usage" 🖱️section, additional clarification was provided for the build and execution process with a command line. This aims to provide clear instructions to users, particularly those who are less familiar with the Go Language.

Remember to keep your project updated, and happy coding! 🎉