## Enhance auto-commit Git alias with git log invocation
In the updated install script, changes have been made to the `auto-commit` Git alias to incorporate a `git log` invocation post auto commit operation.

```diff
-git config --global alias.auto-commit '!git commit -m "$(gh commit)" || git commit -a -m "$(gh commit)"'
+git config --global alias.auto-commit '!git commit -m "$(gh commit)" || git commit -a -m "$(gh commit)" && git log HEAD...HEAD~1'
```

With these changes, after the auto-commit operation, the Git log is invoked to display the most recent commit. This will offer users an immediate overview of their latest commit, including the commit hash, author, date, and message.

This is a small but useful enhancement, making the auto-commit alias not just perform the commit operation, but also presenting commit details right away to the user :+1:.

Please note, however, that the Git log operation will only be executed if the commit is successful, due to the use of the logical AND operator (`&&`). If the commit operation fails, the Git log will not be displayed :warning:.