Title: Added comment about extending functionality using cobra library
## Changes:
- A comment has been added in the main.go file suggesting the use of the Cobra library if more features are needed.

## Details:
The `flag` package is used for basic command-line flag parsing. The code suggests using the `github.com/spf13/cobra` library as it provides a simple interface for creating powerful modern CLI interfaces, and could be helpful when we may want to add more features in the future. This is not a functional change, but rather provides information for potential improvements in codebase and better feature handling.

📝 Please note, this change will not affect the current functionality of the software in any way.