Added environment variable check for diff prompt in util.go
The function `getDiffPrompt` in `util.go` now checks for an environment variable `PROMPT_OVERRIDE`. If this value is not empty, it overrides the default prompt message with the content stored in the environment variable. This adds flexibility to changing the prompt message without needing to modify and redeploy the code. Lastly, the prompt for user input has been changed from "Enter commit message:" to "Commit message:'. :speech_balloon: :wrench: