type: fix: Modify texts in util.go
In this code update, two lines of text in util.go have been changed to enhance clarity and readability. The modification includes the description of how the body of a commit message should be formulated. The new description emphasizes on a beautifully rendered structure instead of just a 'beautiful markdown'.

In addition, the introduction to a commit message has been made more clear by replacing "Commit message:" with "Commit message as follows:". These changes are meant to provide a better user experience in comprehending the instructions and commit messages. :pencil2: :books: