feat: Change trigger of release workflow from repository_dispatch to workflow_dispatch
The 👨‍💻 code of the release workflow was modified 🚧. We had been using `repository_dispatch` as our trigger to kick off the release workflow. However, this method was causing some inconsistencies and unpredictability. 🐛

To improve this, we decided to switch to `workflow_dispatch`. This change 🔄 allows us to manually start the workflow straight from GitHub's interface, offering better control and predictability. 🎮👍
Now, the release workflow will begin with more precision when the tags starting with "v*" are pushed. 🏁🚀 Instantiate the batmobile of seamless deployments! 💨